Use modified stack depth to track record dependencies
Tracking dependencies between records is tricky, so let's not bother.

A simple solution that will work for now is that we start at the root
node with a 'depth' of 0 incrementing and decrementing depending on the
kind of association being traversed.  +1 when the associated object is
dependent on the current one and -1 for the inverse.

Each record then knows its depth which can then be used a priority for
ordering writes.

Many to many is a special case where the associated object is at the
same 'depth' but the join records are one deeper and therefore written
after.

This should ensure that foreign key constraint are never violated always
writing records that are referenced before the records that reference
them.