Hide RelationMapping#factory behind #load method
It seems as though RelationMapping grew out of a value object, so exposing its
factory and serializer in this way was pretty natural. Obviously that isn't
ideal and this same process should be followed for #serializer (e.g. with
a #load method?).

We've reimplemented the test double in the PublicConveniencies spec to reflect
this change, but to do that it's worth noting we have cheated a bit :)