Load new objects into the dirty map
This change ensures that the newly inserted data from new objects that
was not read from the database is put into the dirty map.

Most obviously this solves the issue of a needless writes in the case of
a double save but also keeps the state of the object store consistent
with the database.