Use PostgreSQL upsert for all saves
This took quite a lot more effort than I first thought.

Summary:
- Changed the main upsert method to really do an upsert
- Added a uniqueness index to categories_to_posts for more rigorous tests
- Changed the categories_to_posts foreign key to cascade delete
- Allow a 'Primary key' option in the Cucumber example test
- Emulate duplicate key errors in MemoryAdapter
- Add Sequel upsert API to MemoryAdapter
- `AbstractRecord#non_identity_attributes` is an annoying name, change it
- `AbstractRecord` has a bunch of missing tests, add them

In order to upsert you must know the potentially conflicting fields,
usually the primary key. This caused a bunch of failures in test
scenarios where the primary key or identity of a record isn't known.

There is a tricky situation with the join tables here where they don't
have a primary key but there unique index is identifying so we treat
that as a primary key and end up with empty update fields which causes
Sequel to produce an SQL error `...DO UPDATE SET ` and then there's
nothing to set. In this case update must be set to false.

The change for use with Postgres was not that difficult. I did find the
method `AbstractRecord#non_identity_attributes` annoying and changed it
to `#updatable_attributes` however. Most of the issue was getting the
MemoryAdapter to behave in the same way which makes me think I should
create a true adapter rather than use Sequel directly and have
MemoryAdapter mimic its API.

Finally I found that the Cucumber readme example didn't run, again
because I don't set up the schema with a primary key, so that capability
also had to be added.