Introduce changes API
RelationStore gains two public methods `#changes` and `#changes_sql`
which when called with a modified object graph, returns a representation
of the changes that will be applied.

`#changes` returns an array of record objects, one record object for
each modified row. The record objects contain only the identity /
primary key fields and the changed fields so the delta is more apparent.
Unmodified fields are not included.

`#changes_sql` (when supported) returns the array of SQL statements
that would be applied when saving the modified object graph.

There is one regrettable hack here is a monkey patch the Sequel's
Dataset which is necessary due to a type check it makes when doing a
literal append. See the `SequelDatasetComparisonLiteralAppendPatch`
module and `Sequel::Dataset` for more.