Support for database generated values — IDs, field defaults etc.
The natural behavior of Terrestrial for the application layer to own and
create all data so no values are fed back. This is fine if the application is
set up to generate UUIDs, timestamps etc but many applications cannot
and need support for common database patterns.

This change adds  the following:
* Database generated IDs
  - Persist an object with a nil ID and have it filled in with the
    returned DB value
* Database owned fields such as timestamps generated by triggers
  - These fields are always read and never written.
* Database default values
  - These values will be fed back to objects if the current value is
    `nil`, once set it reverts to normal behavior.
* Library managed timestamps (ActiveRecord style)
  - Terrestrial will manage timestamps for you in much the same way
    ActiveRecord does.