Many-to-many does not assume/need unique index on join table
Closing the door on having a mapped entity in the middle of a
many-to-many association.

To do something like a post's commenters (users that have commented on
it) define a method on `Post` like so

```
class Post
  def commenters
    comments.map(&:commenter)
  end
end
```

This will ensure consistency with comments, not reload comments due to
the comments collection caching but will result in `n+1` queries if
the users are not eager loaded.