Fix dirty-only field insert with non nullable fields
In the dump pipeline the dirty map is used to remove the attributes that
have not changed. This is problematic for upserts in PostreSQL because
a valid insert value for that field must be present in the query.

This hack merges into the attributes a sub-select for each non nullable
field that hasn't been changed.