Allow nil object values in one to many associations
In our example adding a post to a users list of posts causes a foreign
key to written to the post table to point back to the author.

When the post mapping contains a 'belongs to' / 'many to one'
association pointing back to its author then it can be inconvenient to
set up references in both directions.

This change allows a post with a nil author to have the correct author
persisted when the post is added to to the user's posts list.